rename kbc ops and prompt class (#249)
* change names for KBC Pipeline
