+++ dataflow/operators/knowledge_cleaning/generate/kbc_text_cleaner_batch.py
@@ -9,15 +9,19 @@ from dataflow.core import LLMServingABC
 
 
 @OPERATOR_REGISTRY.register()
-class KnowledgeCleanerBatch(OperatorABC):
+class KBCTextCleanerBatch(OperatorABC):
     '''
         KnowledgeCleaner is a class that cleans knowledge for RAG to make them more accurate, reliable and readable.
     '''
 
-    def __init__(self, llm_serving: LLMServingABC, lang="en"):
+    def __init__(self, llm_serving: LLMServingABC, lang="en", prompt_template = None):
         self.logger = get_logger()
-        self.prompts = KnowledgeCleanerPrompt(lang=lang)
+        self.prompts = KnowledgeCleanerPrompt(lang=lang)    
         self.llm_serving = llm_serving
+        if prompt_template:
+            self.prompt_template = prompt_template
+        else:
+            self.prompt_template = KnowledgeCleanerPrompt()
 
     @staticmethod
     def get_desc(lang: str = "zh"):
@@ -61,7 +65,7 @@ class KnowledgeCleanerBatch(OperatorABC):
         Reformat the prompts in the dataframe to generate questions.
         """
         raw_contents = dataframe[self.input_key].tolist()
-        inputs = [self.prompts.Classic_COT_Prompt(
+        inputs = [self.prompt_template.build_prompt(
             raw_content) for raw_content in raw_contents]
 
         return inputs
@@ -89,7 +93,7 @@ class KnowledgeCleanerBatch(OperatorABC):
             raise KeyError("'raw_chunk' field not found in the input file.")
 
         raw_contents = dataframe["raw_chunk"].tolist()
-        inputs = [self.prompts.Classic_COT_Prompt(
+        inputs = [self.prompts.build_prompt(
             raw_content) for raw_content in raw_contents]
 
         return raw_contents, inputs
