+++ dataflow/statics/pipelines/api_pipelines/chemistry_smiles.py
@@ -33,6 +33,42 @@ smiles_prompt = """Extract the monomer/small molecule information from the text 
     Please make sure to output pure json which can be saved into a json file, do not output like html.
     """
 
+response_format = {
+  "type": "json_schema",
+  "json_schema": {
+    "name": "chemical_structures_response",
+    "strict": True,
+    "schema": {
+      "type": "object",
+      "properties": {
+        "chemical_structures": {
+          "type": "array",
+          "items": {
+            "type": "object",
+            "properties": {
+              "abbreviation": {
+                "type": "string"
+              },
+              "full_name": {
+                "type": "string"
+              },
+              "smiles": {
+                "type": "string"
+              }
+            },
+            "required": ["abbreviation", "full_name", "smiles"],
+            "additionalProperties": False
+          }
+        }
+      },
+      "required": ["chemical_structures"],
+      "additionalProperties": False
+    }
+  }
+}
+
+
+
 class ExtractSmiles():
     def __init__(self):
         self.storage = FileStorage(
